refactor and clean up dev server
