+++ lib/socket.ts
@@ -1,0 +1,31 @@
+import http from "http";
+import { Server as IoServer } from "socket.io";
+
+/**
+ * Starts web socket server
+ * @param server Http server
+ * @param callback callback function to handle socket events
+ * @returns 
+ */
+export function startSocket(server: http.Server, callback: (io: IoServer) => void): IoServer {
+  const io = new IoServer(server, {
+    cors: {
+      origin: "*",
+      methods: ["GET", "POST"],
+      allowedHeaders: ["Content-Type", "Authorization"],
+      credentials: true,
+    },
+  });
+
+  log.success("Socket server setup successfully.");
+
+  io.on("connection", (socket) => {
+    log.info(`Socket connected: ${socket.id}`, LogLevels.DEV);
+    socket.on("disconnect", () => {
+      log.info(`Socket disconnected: ${socket.id}`, LogLevels.DEV);
+    })
+    callback(io)
+  });
+
+  return io;
+}
